+++ src/main/java/de/fraunhofer/isst/dataspaceconnector/services/resources/OfferedResourceServiceImpl.java
@@ -12,6 +12,7 @@ import de.fraunhofer.isst.dataspaceconnector.exceptions.resource.InvalidResource
 import de.fraunhofer.isst.dataspaceconnector.exceptions.resource.ResourceAlreadyExistsException;
 import de.fraunhofer.isst.dataspaceconnector.exceptions.resource.ResourceException;
 import de.fraunhofer.isst.dataspaceconnector.exceptions.resource.ResourceNotFoundException;
+import de.fraunhofer.isst.dataspaceconnector.model.BackendSource;
 import de.fraunhofer.isst.dataspaceconnector.model.OfferedResource;
 import de.fraunhofer.isst.dataspaceconnector.model.QueryInput;
 import de.fraunhofer.isst.dataspaceconnector.model.ResourceMetadata;
@@ -381,11 +382,45 @@ public class OfferedResourceServiceImpl implements ResourceService {
     }
 
     /**
+     * Retrieves resource data from the local database by ID or from an external data source using
+     * the given QueryInput.
+     *
+     * @param resourceId ID of the resource
+     * @param queryInput Headers, path variables and params for data request from backend.
+     * @return resource data as string
+     * @throws InvalidResourceException if the resource is invalid.
+     * @throws ResourceNotFoundException if the resource could not be found
+     */
+    public String getData(UUID resourceId, QueryInput queryInput) {
+        final var representations = getAllRepresentations(resourceId);
+        for (var representationId : representations.keySet()) {
+            try {
+                return getDataByRepresentation(resourceId, representationId, queryInput);
+            } catch (ResourceException exception) {
+                // The resource is incomplete or wrong.
+                LOGGER.debug("Resource exception. [resourceId=({}), representationId=({}), exception=({})]", resourceId, representationId, exception);
+            } catch (IllegalArgumentException exception) {
+                // Query input was invalid
+                throw exception;
+            } catch (RuntimeException exception) {
+                // The resource could not be received.
+                LOGGER.debug("Failed to get resource data. [resourceId=({}), representationId=({}), exception=({})]", resourceId, representationId, exception);
+            }
+        }
+
+        // This code should never be reached since the representation should have at least one
+        // representation.
+        invalidResourceGuard(getResource(resourceId));
+        // Add a runtime exception in case the resource valid logic changed.
+        throw new RuntimeException("This code should not have been reached.");
+    }
+
+    /**
      * Retrieves resource data from the local database or an external data source by ID.
      *
      * @param resourceId ID of the resource
      * @param representationId ID of the representation
-     * @param queryInput Header and params for data request from backend.
+     * @param queryInput Headers, path variables and params for data request from backend.
      * @return resource data as string
      * @throws ResourceNotFoundException if the resource could not be found
      * @throws ResourceException if the resource data could not be retrieved
@@ -537,6 +572,24 @@ public class OfferedResourceServiceImpl implements ResourceService {
             return Optional.of("The resource representation cannot be null.");
         }
 
+        for (ResourceRepresentation representation :
+                resource.getResourceMetadata().getRepresentations().values()) {
+            BackendSource source = representation.getSource();
+            if (source.getType().equals(BackendSource.Type.HTTP_GET)
+                    || source.getType().equals(BackendSource.Type.HTTPS_GET)
+                    || source.getType().equals(BackendSource.Type.HTTPS_GET_BASICAUTH)) {
+                long openingBracesCount = source.getUrl().toString().chars()
+                        .filter(ch -> ch == '{').count();
+                long closingBracesCount = source.getUrl().toString().chars()
+                        .filter(ch -> ch == '}').count();
+
+                if (openingBracesCount != closingBracesCount) {
+                    return Optional.of("URL of backend source must contain same number of"
+                            + " '{' and '}'");
+                }
+            }
+        }
+
         return Optional.empty();
     }
 
@@ -606,6 +659,9 @@ public class OfferedResourceServiceImpl implements ResourceService {
                         resource, representation, exception);
                 throw new ResourceException("The deposited address is not a valid URI.",
                     exception);
+            } catch (IllegalArgumentException exception) {
+                // Query input was invalid
+                throw exception;
             } catch (RuntimeException exception) {
                 // One of the http calls encountered problems.
                 LOGGER.debug("Failed to establish source connection. [resource=({}), " +